Try to fix Windows Ci

+++ src/joystick/hidapi/SDL_hidapi_psmove.c
@@ -220,34 +220,34 @@ static void HIDAPI_DriverPSMove_HandleStatePacket(SDL_Joystick *joystick, SDL_Dr
 
        float sensor_data[3];
         if(ctx->model == Model_ZCM1) {
-            sensor_data[0] = (((ctx->input.common.aX[0] + ctx->input.common.aX2[0]) +
+            sensor_data[0] = (float)(((ctx->input.common.aX[0] + ctx->input.common.aX2[0]) +
                    ((ctx->input.common.aX[1] + ctx->input.common.aX2[1]) << 8)) / 2 - 0x8000);
 
-            sensor_data[1] = -(((ctx->input.common.aY[0] + ctx->input.common.aY2[0]) +
+            sensor_data[1] = -(float)(((ctx->input.common.aY[0] + ctx->input.common.aY2[0]) +
                    ((ctx->input.common.aY[1] + ctx->input.common.aY2[1]) << 8)) / 2 - 0x8000);
 
-            sensor_data[2] = -(((ctx->input.common.aZ[0] + ctx->input.common.aZ2[0]) +
+            sensor_data[2] = -(float)(((ctx->input.common.aZ[0] + ctx->input.common.aZ2[0]) +
                    ((ctx->input.common.aZ[1] + ctx->input.common.aZ2[1]) << 8)) / 2 - 0x8000);
         } else {
-            sensor_data[0] = (Sint16)(ctx->input.common.aX[0] + (ctx->input.common.aX[1] << 8));
-            sensor_data[1] = (Sint16)((ctx->input.common.aY[0] + (ctx->input.common.aY[1] << 8)));
-            sensor_data[2] = (Sint16)((ctx->input.common.aZ[0] + (ctx->input.common.aZ[1] << 8)));
+            sensor_data[0] = (float)(Sint16)(ctx->input.common.aX[0] + (ctx->input.common.aX[1] << 8));
+            sensor_data[1] = (float)(Sint16)((ctx->input.common.aY[0] + (ctx->input.common.aY[1] << 8)));
+            sensor_data[2] = (float)(Sint16)((ctx->input.common.aZ[0] + (ctx->input.common.aZ[1] << 8)));
         }
         SDL_SendJoystickSensor(timestamp, joystick, SDL_SENSOR_ACCEL, timestamp, sensor_data, SDL_arraysize(sensor_data));
 
         if(ctx->model == Model_ZCM1) {
-            sensor_data[0] = (((ctx->input.common.gX[0] + ctx->input.common.gX2[0]) +
+            sensor_data[0] = (float)(((ctx->input.common.gX[0] + ctx->input.common.gX2[0]) +
                    ((ctx->input.common.gX[1] + ctx->input.common.gX2[1]) << 8)) / 2 - 0x8000);
 
-            sensor_data[1] = (((ctx->input.common.gY[0] + ctx->input.common.gY2[0]) +
+            sensor_data[1] = (float)(((ctx->input.common.gY[0] + ctx->input.common.gY2[0]) +
                    ((ctx->input.common.gY[1] + ctx->input.common.gY2[1]) << 8)) / 2 - 0x8000);
 
-            sensor_data[2] = (((ctx->input.common.gZ[0] + ctx->input.common.gZ2[0]) +
+            sensor_data[2] = (float)(((ctx->input.common.gZ[0] + ctx->input.common.gZ2[0]) +
                    ((ctx->input.common.gZ[1] + ctx->input.common.gZ2[1]) << 8)) / 2 - 0x8000);
         } else {
-            sensor_data[0] = (Sint16)(ctx->input.common.gX[0] + (ctx->input.common.gX[1] << 8));
-            sensor_data[1] = (Sint16)(ctx->input.common.gY[0] + (ctx->input.common.gY[1] << 8));
-            sensor_data[2] = (Sint16)(ctx->input.common.gZ[0] + (ctx->input.common.gZ[1] << 8));
+            sensor_data[0] = (float)(Sint16)(ctx->input.common.gX[0] + (ctx->input.common.gX[1] << 8));
+            sensor_data[1] = (float)(Sint16)(ctx->input.common.gY[0] + (ctx->input.common.gY[1] << 8));
+            sensor_data[2] = (float)(Sint16)(ctx->input.common.gZ[0] + (ctx->input.common.gZ[1] << 8));
         }
         SDL_SendJoystickSensor(timestamp, joystick, SDL_SENSOR_GYRO, timestamp, sensor_data, SDL_arraysize(sensor_data));
     }
